src/analytics/analytics_module.py


# analytics_module.py

import datetime
import pandas as pd  # Предположим, что мы используем pandas для обработки данных

class Analytics:
    def __init__(self):
        # Здесь может быть код инициализации, например, подключение к базе данных
        pass

    def calculate_average(self, data):
        """
        Вычисляет среднее значение списка.
        """
        if not data:
            return 0
        return sum(data) / len(data)

    def generate_report(self, data):
        """
        Генерирует аналитический отчет на основе предоставленных данных.
        """
        df = pd.DataFrame(data)
        report = {
            "average_value": self.calculate_average(df["value"].tolist()),
            "max_value": df["value"].max(),
            "min_value": df["value"].min(),
            "data_count": len(df)
        }
        return report

    def daily_trend_analysis(self, data):
        """
        Анализирует тренды на ежедневной основе.
        """
        df = pd.DataFrame(data)
        df['date'] = pd.to_datetime(df['date'])
        df.set_index('date', inplace=True)
        daily_trends = df.resample('D').mean()
        return daily_trends.to_dict()

    def monthly_performance(self, data):
        """
        Анализирует ежемесячную производительность.
        """
        df = pd.DataFrame(data)
        df['date'] = pd.to_datetime(df['date'])
        df.set_index('date', inplace=True)
        monthly_performance = df.resample('M').sum()
        return monthly_performance.to_dict()

if __name__ == "__main__":
    analytics = Analytics()
    
    # Пример данных
    sample_data = [
        {"date": "2023-01-01", "value": 100},
        {"date": "2023-01-02", "value": 105},
        {"date": "2023-01-03", "value": 103},
        # ... и так далее
    ]

    report = analytics.generate_report(sample_data)
    print("Report:", report)

    trends = analytics.daily_trend_analysis(sample_data)
    print("Daily Trends:", trends)
