requirements.txt
# Websockets and async communication
websockets==10.0

# AWS services
boto3==1.20.23
botocore==1.23.23

# Slack integration
slack-sdk==3.11.2

# Database (DynamoDB)
pynamodb==5.2.0

# Testing
pytest==7.2.1
pytest-mock==3.7.3

# For API and HTTP requests
requests==2.26.0

# Utils and other necessary libraries
python-dotenv==0.19.2

# YAML parser
pyyaml==5.4.1

# OpenAI SDK
openai==0.27.3

#npm install -g serverless